requirements.txt
# GUI框架
tkinter-tooltip==3.0.0

# HTTP客户端
requests==2.31.0
aiohttp==3.9.1  # 异步HTTP支持

# 图片处理
Pillow==10.1.0
pdf2image==1.16.3
PyMuPDF==1.26.3  # 更好的PDF处理

# PPT处理
python-pptx==1.0.2  # 升级到最新版本
# spire.presentation  # 商业库，需要单独安装

# 配置管理
configparser==6.0.0

# 工具库
tqdm==4.66.1
colorama==0.4.6

# 测试框架
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# 代码质量
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# 日志和监控
loguru==0.7.2

# 类型提示
typing-extensions==4.9.0

# 文档转换（可选）
pypandoc==1.12  # 用于转换文档格式
unoconv==0.9.0  # LibreOffice命令行工具

# 开发工具（可选）
ipython==8.18.1
jupyter==1.0.0